Also move IMigratableModel into migrator directory
