--- examples/utils/migration-tools/src/interfaces/index.ts
@@ -1,10 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export {
-	IImportExportModel,
-	IMigratableModel,
-	IVersionedModel,
-} from "./migratableModel.js";
